requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
requests==2.32.3
python-decouple==3.8
reportlab==4.2.5
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
